+++ documentation/src/components/Layout.astro
@@ -1,4 +1,6 @@
 ---
+import SearchMenu from "./SearchMenu.astro";
+
 interface Props {
 	meta: {
 		title: string | null;
@@ -23,19 +25,6 @@ const origin = import.meta.env.DEV
 	? Astro.url.origin
 	: "https://lucia-auth.com";
 ---
-
-<script>
-	import { addClassName, removeClassName } from "@lib/dom";
-	import { onMenuToggle } from "@lib/state";
-
-	onMenuToggle((isMenuOpen) => {
-		if (isMenuOpen) {
-			addClassName(document.body, "overflow-hidden");
-		} else {
-			removeClassName(document.body, "overflow-hidden");
-		}
-	});
-</script>
 
 <html lang="en" class="bg-main">
 	<head>
@@ -148,14 +137,12 @@ const origin = import.meta.env.DEV
 	</head>
 
 	<body
-		class:list={[
-			"dark:text-zinc-200 leading-relaxed text-shadow-zinc xl:overflow-auto",
-			theme
-		]}
+		class:list={["dark:text-zinc-200 leading-relaxed text-shadow-zinc", theme]}
 		set:html={slotHtml}
 	>
 		<!--not super proud of this - need this to hide main-color bg on chrome -->
 		<div class="fixed top-0 -z-10 h-screen w-screen bg-white dark:bg-zinc-900">
 		</div>
+		<SearchMenu />
 	</body>
 </html>
