+++ documentation/src/components/SearchMenu.astro
@@ -1,0 +1,61 @@
+---
+import SearchBar from "./search/SearchBar";
+---
+
+<div
+	class="text-shadow-zinc fixed z-50 hidden h-full w-full place-content-center place-items-center overscroll-contain bg-black bg-opacity-20 py-36 px-4 leading-relaxed dark:bg-white dark:bg-opacity-5 dark:text-zinc-200"
+	id="search-backdrop"
+>
+	<div
+		class="dark:bg-shadow-zinc h-full w-full max-w-xl rounded-xl bg-white px-6 py-8 shadow-xl"
+		id="search-modal"
+	>
+		<SearchBar client:load />
+	</div>
+</div>
+
+<script>
+	import {
+		addClassName,
+		removeClassName,
+		freezePage,
+		unFreezePage
+	} from "@lib/dom";
+	import { onSearchMenuStateUpdate, setSearchMenuState } from "@lib/state";
+
+	import { focusOnSearchBar } from "./search/SearchBar";
+
+	const searchBackdropElement = document.getElementById("search-backdrop");
+	const searchModalElement = document.getElementById("search-modal");
+
+	if (
+		searchBackdropElement instanceof HTMLDivElement &&
+		searchModalElement instanceof HTMLDivElement
+	) {
+		const closeSearchMenu = () => {
+			addClassName(searchBackdropElement, "hidden");
+			removeClassName(searchBackdropElement, "flex");
+			unFreezePage();
+
+		};
+
+		const openSearchMenu = () => {
+			removeClassName(searchBackdropElement, "hidden");
+			addClassName(searchBackdropElement, "flex");
+			freezePage();
+            focusOnSearchBar();
+		};
+
+		searchBackdropElement.addEventListener("click", (e) => {
+			setSearchMenuState(false);
+		});
+		searchModalElement.addEventListener("click", (e) => {
+			e.stopPropagation();
+		});
+
+		onSearchMenuStateUpdate((searchMenuOpen) => {
+			if (searchMenuOpen) return openSearchMenu();
+			closeSearchMenu();
+		});
+	}
+</script>
