+++ documentation/src/lib/state.ts
@@ -1,16 +1,20 @@
-const createToggleStore = (initialValue = false) => {
+const createStore = <V>(initialValue: V) => {
 	let value = initialValue;
-	const subscribers: ((e: boolean) => void)[] = [];
-	const subscribe = (callback: (e: boolean) => void) => {
+	const subscribers: ((e: V) => void)[] = [];
+	const subscribe = (callback: (e: V) => void) => {
 		subscribers.push(callback);
 	};
-	const update = () => {
-		value = !value;
+	const set = (newValue: V | ((val: V) => V)) => {
+		if (newValue instanceof Function) {
+			newValue = newValue(value);
+		}
+		value = newValue;
 		for (const subscriber of subscribers) {
 			subscriber(value);
 		}
 	};
-	return [subscribe, update] as const;
+	return [subscribe, set] as const;
 };
 
-export const [onMenuToggle, toggleMenu] = createToggleStore();
+export const [onMenuStateUpdate, setMenuState] = createStore(false);
+export const [onSearchMenuStateUpdate, setSearchMenuState] = createStore(false);
