+++ documentation/src/pages/[...content].astro
@@ -125,3 +125,12 @@ const contributePageHref = joinUrlPaths(
 	</div>
 	<Banner />
 </Layout>
+
+<script>
+	// const frameworkIdElement = document.getElementById("framework-id");
+	// if (frameworkIdElement instanceof HTMLInputElement) {
+	// 	const url = new URL(window.location.href);
+	// 	url.searchParams.set("framework", frameworkIdElement.value || "none");
+	// 	window.history.replaceState({}, "", url);
+	// }
+</script>
